Revised the structure of counting model
