+++ datastructs/counters/lossy_counter/lossycon.c
@@ -8,6 +8,10 @@
 #include <inttypes.h>
 #include <string.h>
 #include <arpa/inet.h>
+#include <time.h>
+#include <math.h>
+
+#include "lossycon.h"
 /****************************************
         Author: Tim Wood
                 Chenghu He
@@ -25,34 +29,6 @@
         to Creative Commons, 559 Nathan Abbott Way, Stanford, California
         94305, USA. 
 ****************************************/
-
-#define BACKLOG 10     // how many pending connections queue will hold
-
-/* for lossy counting */
-typedef struct counter
-{
-        int item;
-        int count;
-} Counter;
-
-typedef struct LC_type
-{
-        Counter *bucket;
-        Counter *holder;
-        Counter *newcount;
-        int buckets;
-        int holdersize;
-        int maxholder;
-        int window;
-        int epoch;
-} LC_type;
-
-// conter functions
-LC_type * LC_Init(float);
-void LC_Delete(LC_type *);
-void LC_Update(LC_type *, int);
-int LC_Size(LC_type *, int);
-Counter * LC_Output(LC_type *,int);
 
 // LC implementation
 LC_type * LC_Init(float phi)
@@ -372,41 +348,3 @@ int recv_key_tcp(char *server_port, char *lossy_phi)
         return 0;
 }
 
-int main(int argc, char ** argv)
-{
-        char* server_port = "1234";
-        char* lossy_phi = "0.01";
-        int o;
-
-        /* Command line args:
-                -p port
-                -i phi for lossy counting
-        */
-        while ((o = getopt (argc, argv, "p:")) != -1) {
-                switch(o){
-                case 'p':
-                        server_port = optarg;
-                        break;
-                case 'i':
-                        lossy_phi = optarg;
-                        break;
-                case '?':
-                        if(optopt == 'p' || optopt == 'i') {
-                                fprintf (stderr, "Option %c requires an argument.\n", optopt);
-                        }
-                        else {
-                                fprintf (stderr, "Unknown argument: %c.\n", optopt);
-                        }
-                        break;
-                }
-        }
-
-        printf("Lossy counter usage: -p [port] -i [phi]\n");
-
-        if (recv_key_tcp(server_port, lossy_phi) != 0) {
-                perror("Error on recv key by tcp");
-                exit(-1);
-        }
-
-        return 0;
-}
