--- datastructs/counters/zipf_generator/Makefile
@@ -1,9 +1,0 @@
-default:all
-
-all:zipfgen
-
-zipfgen:zipfgen.c
-	gcc -o zipfgen zipfgen.c
-
-clean:
-	rm zipfgen
